new solution

+++ solutions/src/countSquareSumTriples/CountSquareSumTriples.java
@@ -5,7 +5,7 @@ public class CountSquareSumTriples {
         int ans = 0;
         for (int a = 1; a <= n; a++) {
             for (int b = 1; b <= n; b++) {
-                double prod = (double) a * a + b * b;
+                double prod = a * a + b * b;
                 int c = (int) Math.sqrt(prod);
                 if (c <= n && c * c == prod) {
                     ans++;
